feat: introduce TipoUsuario enum
